+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - `Persistent.equals`, `Persistent.notEquals`, `Persistent.assign` now support function evaluator as argument
 - Added `hidden` and `disabled` config to `Menu` choice
 - Added `Menu.hideIf` and `Menu.disableIf` magic methods
+- Added `Menu.enableWhen` and `Menu.showWhen`
 
 ## [0.8.2]